Removes a bunch of duplication in the `msg.Len` path and exposes the method used to calculate the compressed length of a RR given a provided compression map.