@rolandshoemaker 2017-09-21
```
master HEAD
BenchmarkMsgLength-4                	 1000000	      1576 ns/op
BenchmarkMsgLengthNoCompression-4   	20000000	       108 ns/op

compress-len HEAD
BenchmarkMsgLength-4                	 1000000	      1552 ns/op
BenchmarkMsgLengthNoCompression-4   	10000000	       123 ns/op
```

Appears to have minimal impact.

Reason for exporting CompressLen was to help work on a external implementation of a fit message in X bytes method although I now notice there is an issue open for a similar method here (#223). I'd be happy to leave this unexported and instead just upstream the rest of my work on that here (I feel like there could be utility for leaving it exported but I can't really come up with another good reason off the top of my head...).