@miekg 2017-09-21
From 108ns/op to 123 ns/op that