@rolandshoemaker 2017-09-21
So I think the most of the latency increase here comes from using a nested loop that the compiler can't optimize as much as it'd like. If I pull that out we get down to ~9ns difference, although this is unlikely to be a statistically significant change (`0.063 > 0.05`)

```
name                      old time/op  new time/op  delta
MsgLength-4               1.79µs ± 4%  1.92µs ±15%   ~     (p=0.310 n=5+5)
MsgLengthNoCompression-4   126ns ± 8%   135ns ± 7%   ~     (p=0.063 n=5+5)
```

but ¯\_(ツ)_/¯, if you're not happy with any increase here I'm happy to just close this out.