@miekg 2017-09-26
No, we should not merge anything that increases time spend here. If you can get a lower I'll merge :)

Closing for now.